keep params in runtime common
