+++ runtime/common/src/lib.rs
@@ -25,7 +25,9 @@ pub mod migration;
 pub mod test_helpers;
 
 use frame_support::{parameter_types, weights::Weight};
-use manta_primitives::constants::WEIGHT_PER_NANOS;
+use manta_primitives::{constants::WEIGHT_PER_NANOS, types::BlockNumber};
+use pallet_transaction_payment::{Multiplier, TargetedFeeAdjustment};
+use sp_runtime::{traits::Bounded, FixedPointNumber, Perquintill};
 
 // From https://github.com/paritytech/polkadot/pull/4332/files?diff=unified&w=1 @ runtime/common/src/lib.rs
 /// Macro to set a value (e.g. when using the `parameter_types` macro) to either a production value
@@ -59,6 +61,31 @@ macro_rules! prod_or_fast {
         }
     };
 }
+
+parameter_types! {
+    pub const BlockHashCount: BlockNumber = 2400;
+    /// The portion of the `NORMAL_DISPATCH_RATIO` that we adjust the fees with. Blocks filled less
+    /// than this will decrease the weight and more will increase.
+    pub const TargetBlockFullness: Perquintill = Perquintill::from_percent(25);
+    /// The adjustment variable of the runtime. Higher values will cause `TargetBlockFullness` to
+    /// change the fees more rapidly.
+    pub AdjustmentVariable: Multiplier = Multiplier::saturating_from_rational(225, 100_000);
+    /// Minimum amount of the multiplier. This value cannot be too low. A test case should ensure
+    /// that combined with `AdjustmentVariable`, we can recover from the minimum.
+    /// See `multiplier_can_grow_from_zero`.
+    pub MinimumMultiplier: Multiplier = Multiplier::saturating_from_rational(1, 5_000u128);
+    pub MaximumMultiplier: Multiplier = Bounded::max_value();
+}
+
+/// Parameterized slow adjusting fee updated based on
+/// https://research.web3.foundation/en/latest/polkadot/overview/2-token-economics.html#-2.-slow-adjusting-mechanism
+pub type SlowAdjustingFeeUpdate<R> = TargetedFeeAdjustment<
+    R,
+    TargetBlockFullness,
+    AdjustmentVariable,
+    MinimumMultiplier,
+    MaximumMultiplier,
+>;
 
 parameter_types! {
     /// Time to execute an empty block.
