+++ runtime/manta/src/lib.rs
@@ -66,7 +66,9 @@ use manta_primitives::{
 };
 pub use pallet_parachain_staking::{InflationInfo, Range};
 use pallet_session::ShouldEndSession;
-use runtime_common::{prod_or_fast, BlockExecutionWeight, ExtrinsicBaseWeight};
+use runtime_common::{
+    prod_or_fast, BlockExecutionWeight, BlockHashCount, ExtrinsicBaseWeight, SlowAdjustingFeeUpdate,
+};
 use session_key_primitives::{AuraId, NimbusId, VrfId};
 
 #[cfg(feature = "runtime-benchmarks")]
@@ -84,7 +86,6 @@ pub mod staking;
 pub mod xcm_config;
 
 use currency::*;
-use fee::{BlockHashCount, SlowAdjustingFeeUpdate};
 use impls::DealWithFees;
 
 pub type NegativeImbalance = <Balances as Currency<AccountId>>::NegativeImbalance;
